requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
greenlet==3.2.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
qrcode[pil]==7.4.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.0.1
flask-login
flask-dance
flask-wtf
email_validator
requests==2.31.0
googletrans==4.0.0rc1
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
Flask-Mail>=0.9.1